// src/models/V1NodeSystemInfo/types.ts
export interface V1NodeSystemInfoModelOptions {
  architecture?: string;
  bootID?: string;
  containerRuntimeVersion?: string;
  kernelVersion?: string;
  kubeProxyVersion?: string;
  kubeletVersion?: string;
  machineID?: string;
  operatingSystem?: string;
  osImage?: string;
  systemUUID?: string;
}